Refactor another test
